chore: move registration mutation hook
- separated schemas
- renamed hooks extension
