+++ packages/apps/human-app/frontend/src/modules/worker/oracle-registration/hooks/use-exchange-oracle-registration-mutation.ts
@@ -1,24 +1,8 @@
-/* eslint-disable camelcase */
 import { z } from 'zod';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { ethers } from 'ethers';
-import { t } from 'i18next';
 import { apiClient } from '@/api/api-client';
 import { apiPaths } from '@/api/api-paths';
-
-export const registrationInExchangeOracleDtoSchema = z.object({
-  oracle_address: z
-    .string()
-    .refine(
-      (address) => ethers.isAddress(address),
-      t('validation.invalidOracleAddress')
-    ),
-  h_captcha_token: z.string().min(1, t('validation.captcha')).default('token'),
-});
-
-export type RegistrationInExchangeOracleDto = z.infer<
-  typeof registrationInExchangeOracleDtoSchema
->;
+import { type RegistrationInExchangeOracleDto } from '../schema';
 
 const RegistrationInExchangeOracleSuccessResponseSchema = z.unknown();
