+++ packages/apps/human-app/frontend/src/modules/worker/oracle-registration/hooks/use-oracle-instructions.ts
