+++ packages/apps/human-app/frontend/src/modules/worker/oracle-registration/schema.ts
@@ -1,0 +1,18 @@
+/* eslint-disable camelcase */
+import { ethers } from 'ethers';
+import { t } from 'i18next';
+import { z } from 'zod';
+
+export const registrationInExchangeOracleDtoSchema = z.object({
+  oracle_address: z
+    .string()
+    .refine(
+      (address) => ethers.isAddress(address),
+      t('validation.invalidOracleAddress')
+    ),
+  h_captcha_token: z.string().min(1, t('validation.captcha')).default('token'),
+});
+
+export type RegistrationInExchangeOracleDto = z.infer<
+  typeof registrationInExchangeOracleDtoSchema
+>;
