--- packages/apps/human-app/frontend/src/modules/worker/oracle-registration/sevices/index.ts
@@ -1,1 +1,0 @@
-export * from './registration-in-exchange-oracles';
